Praktické cvičení_lekce7/requirements.txt
langchain
langchain-community
langchain-experimental
langchain-openai
openai
python-dotenv
requests
sqlalchemy
numpy